Convert ModalSmall to ES6 component
